refactor: move expandable file path
