+++ src/components/common/utilities/expandable.tsx
